feat(workflow_runs): Add Workflow Run as new schema and stream choice
