+++ tap_github/gh_client.py
@@ -2,12 +2,10 @@
 import argparse
 import os
 import json
-import collections
 import time
 import requests
 # Third-Party imports
 import singer
-import singer.bookmarks as bookmarks
 import singer.metrics as metrics
 import backoff
 # Project imports
@@ -17,58 +15,6 @@ from exceptions import *
 REQUEST_TIMEOUT = 300
 session = requests.Session()
 logger = singer.get_logger()
-
-
-def get_bookmark(state, repo, stream_name, bookmark_key, start_date):
-    repo_stream_dict = bookmarks.get_bookmark(state, repo, stream_name)
-    if repo_stream_dict:
-        return repo_stream_dict.get(bookmark_key)
-    if start_date:
-        return start_date
-    return None
-
-
-def translate_state(state, catalog, repositories):
-    '''
-    This tap used to only support a single repository, in which case the
-    state took the shape of:
-    {
-      "bookmarks": {
-        "commits": {
-          "since": "2018-11-14T13:21:20.700360Z"
-        }
-      }
-    }
-    The tap now supports multiple repos, so this function should be called
-    at the beginning of each run to ensure the state is translate to the
-    new format:
-    {
-      "bookmarks": {
-        "singer-io/tap-adwords": {
-          "commits": {
-            "since": "2018-11-14T13:21:20.700360Z"
-          }
-        }
-        "singer-io/tap-salesforce": {
-          "commits": {
-            "since": "2018-11-14T13:21:20.700360Z"
-          }
-        }
-      }
-    }
-    '''
-    nested_dict = lambda: collections.defaultdict(nested_dict)
-    new_state = nested_dict()
-
-    for stream in catalog['streams']:
-        stream_name = stream['tap_stream_id']
-        for repo in repositories:
-            if bookmarks.get_bookmark(state, repo, stream_name):
-                return state
-            if bookmarks.get_bookmark(state, stream_name, 'since'):
-                new_state['bookmarks'][repo][stream_name]['since'] = bookmarks.get_bookmark(state, stream_name, 'since')
-
-    return new_state
 
 
 def raise_for_error(resp, source):
